pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "aumai-specs"
version = "1.0.0"
description = "Canonical schema definitions and Pydantic models for all AumAI projects"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "MuVeraAI Corporation", email = "engineering@muveraai.com" },
]
keywords = ["aumai", "agents", "schemas", "llm", "ai", "specifications"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Typing :: Typed",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "jsonschema>=4.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "mypy>=1.10",
    "ruff>=0.4",
    "pre-commit>=3.0",
]

[project.urls]
Homepage = "https://github.com/muveraai/aumai-specs"
Documentation = "https://specs.aumai.dev"
Repository = "https://github.com/muveraai/aumai-specs"
Issues = "https://github.com/muveraai/aumai-specs/issues"
Changelog = "https://github.com/muveraai/aumai-specs/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/aumai_specs"]

[tool.hatch.build.targets.wheel.sources]
"src" = ""

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--strict-markers",
    "--tb=short",
    "-v",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests requiring external services",
]

# ---------------------------------------------------------------------------
# coverage
# ---------------------------------------------------------------------------

[tool.coverage.run]
source = ["src/aumai_specs"]
omit = ["*/py.typed"]
branch = true

[tool.coverage.report]
show_missing = true
skip_empty = true
fail_under = 90
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@overload",
    "\\.\\.\\.",
]

# ---------------------------------------------------------------------------
# mypy
# ---------------------------------------------------------------------------

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
show_error_codes = true
pretty = true
files = ["src/aumai_specs"]
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "jsonschema.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "yaml"
ignore_missing_imports = true

# ---------------------------------------------------------------------------
# ruff (linting + formatting)
# ---------------------------------------------------------------------------

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "N",    # pep8-naming
    "ANN",  # flake8-annotations
    "S",    # flake8-bandit (security)
    "T20",  # flake8-print
    "RUF",  # ruff-specific rules
]
ignore = [
    "ANN101",  # Missing type annotation for `self`
    "ANN102",  # Missing type annotation for `cls`
    "S101",    # Use of assert — fine in tests
]

[tool.ruff.lint.per-file-ignores]
"tests/**" = ["ANN", "S", "T20"]
"examples/**" = ["ANN", "T20"]

[tool.ruff.lint.isort]
known-first-party = ["aumai_specs"]
